metrics: add failed task metrics to collector
