+++ server/impl/src/test/java/com/walmartlabs/concord/server/tasks/SchedulerDaoTest.java
@@ -1,0 +1,48 @@
+package com.walmartlabs.concord.server.tasks;
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import com.walmartlabs.concord.server.AbstractDaoTest;
+import com.walmartlabs.concord.server.metrics.FailedTaskError;
+import com.walmartlabs.concord.server.task.SchedulerDao;
+import org.junit.Before;
+import org.junit.Ignore;
+import org.junit.Test;
+
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+
+@Ignore("requires a local DB instance")
+public class SchedulerDaoTest extends AbstractDaoTest {
+
+    private SchedulerDao schedulerDao;
+
+    @Before
+    public void setUp() {
+        schedulerDao = new SchedulerDao(getConfiguration());
+    }
+
+    @Test
+    public void testPollErrored() {
+        List<FailedTaskError> failedTaskErrors = schedulerDao.pollErrored();
+        assertEquals(0, failedTaskErrors.size());
+    }
+}
